Add servlet-api to the maven POM. This is duplicated with ScalazProject.scala. An alternative would be to make the POM the canonical source of dependencies for SBT. Given that we we only have a handful of dependencies the duplication is acceptable.
Give the POM a non-default name to ensure that SBT does not consider it as a source of dependency config.

I've got a feeling that if any of any class from one of our optional dependencies (ScalaCheck / servlet-api / Functional Java) appears in a signature in Scalaz.scala, they become mandatory compile-time dependencies for a user of Scalaz. Specs tries to solve this problem by providing a set of traits that you can mix and max into a base test class. Perhaps we could do something similar.

git-svn-id: https://scalaz.googlecode.com/svn/trunk@750 ea79b6a6-c355-11dd-8bad-cde060d8fd3c