Throw a bit more by-name-ness in Apply.
This enables partial Traverse, given a suitably lazy Applicative.

Stream.from(1).traverseU((x: Int) => if (x < 3) // None

Currently only Option[Applicative] is known to exhibit this laziness. Others should be reviewed. It might make sense to make the `fa` parameter in `Apply#apply` strict once again, so it's clear to the implementer of the method which order to evaluate the arguments.