Identify Equal[T] instances based on Any#==. ("natural" equality)
This can be used for fast-path equality checks, e.g. for Sets.

Equal[Option[A]] is natural if Equal[A] is natural.

Would be interesting to expose this property as a type member, rather than a value member, of Equal, but that's one for another day.

Laws for Lens. Instance for Set.