Added whenM/unlessM to BooleanFunctions and BooleanOps.
These implementations are equiv to Control.Monad.when and Control.Monad.unless. The
names use the M suffix as syntax already exists for when/unless. The existing
implementations are used only for side-effects. Optimally it would be nicer to
special case the side-effecting functions, but existing programs would continue
to compile even though they are broken if when/unless were reused.