Remove Pointed, Copointed and ArrId. The existence of these three type-classes is a bug #235.
The operations of these former interfaces have been moved to a more appropriate place:
* Pointed operation has been moved to Applicative
* Copointed operation has been moved to Comonad
* ArrId operation has been moved to Category

A short discussion occurred on the mailing list on this issue:
https://groups.google.com/forum/?fromgroups=#!topic/scalaz/7OE_Nsreqq0

Pointed is a bogus type-class for a number of reasons. The first is that it provides no value.
It's only possible law is given by paramericity (c.f. Theorems for Free, Wadler):
    (x, g) => point(map(x)(g)) == pure(g(x))

Copointed and ArrId have similar reasoning applied.

The adversities of these interfaces have long been discussed; most notably between Ed Kmett, Runar Bjarnason and I.
There is consensus that their existence serves no purpose but to attract defective code. This is the primary compulsion for their removal (not so much their uselessness). This bug fix is similar in structure to the removal of the Zero interface (scalaz 6) with the zero operation moved to the Monoid interface. I cannot find the original discussion on this matter, but I do recall that there were more people involved and also unanimous consensus.

Since I was the original author of these unfortunate type-classes, I offer my apologies for their coming to exist. I co-invented these type-classes with Ed (After I had invented them, I looked around for names and found Ed doing the same thing and stole his names!). I hope my apology can be accepted and we can move on.