InsertionMap equality should be order independent.
The underlying structure stores an index that _can_
be used to pull keys back out in the insertion order,
but this index is not part of data and should not be
taken into account for equal / hashCode and alike.

This fits with the rest of the data structure, for
example map/flatMap occur in an arbitrary order.

This reverts a behaviour change introduced in
commit ec854db7180dbba78d19f6df6a0c040034f3245b