Remove parsing on Scala version.
I added this when I was bridging 2.11.0-M4/-M5, the time
that we modularised scala-xml and scala-parser-combinators.

Now, we just assume that "2.11.*" is modularized.

Nice side effect: the build is not tolerant of the custom,
SHA-based scala versions used in integration builds with
dbuild [1].

This change should be merged to scala-seven

[1] http://typesafehub.github.io/distributed-build/0.7.0/index.html