Improve functions in Map.scala by adding balance{L, R} functions
As balance function checks left and right subtrees' balance, which is unnecessary for some cases,
we can improve this redundancy by introducing balanceL and balanceR functions.[1,2,3,4]

Note that alter function still needs balance function as there might be a case that we insert or
delete a key-value from a subtree and balance the subtree afterwords.
So we can't remove balance function.

This patch modified the functions below:
  - join
  - merge
  - glue
  - insert
  - insertWithKey
  - delete
  - updateWithKey
  - updateLookupWithKey
  - updateAt
  - deleteMax
  - deleteMin
  - updateMinWithKey
  - insertMax
  - insertMin
  - deleteFindMax
  - deleteFindMin

[1] https://github.com/haskell/containers/commit/4193a09336772748bdec0cdab5da4257d224c814
[2] https://github.com/haskell/containers/commit/1262e1d607aaf0f3a21626ff2e9810c070fa70eb#diff-203612a31f8727a5e95289893d980b9dL1894
[3] https://mail.haskell.org/pipermail/libraries/2010-September/014254.html
[4] https://mail.haskell.org/pipermail/libraries/2010-September/014343.html