Unambiguous ops (syntax extensions).
The goal is to allow the user to structure imports so that ambiguous
syntactic extensions are avoided.

The idea is that the import

  import scalaz.syntax.traverse0._

imports syntax for Traverse, but doesn't include syntax for Functor and Foldable.
The usual import

  import scalaz.syntax.traverse._

also brings in the syntax for Functor and Foldable.
There is one change, though: for this import to provide Functor (or Foldable)
syntax, there needs to be a Traverse instance in scope; Functor (or Foldable)
instance will not suffice, even when only Functor (or Foldable) operations
are performed.

This allows us to import

  import scalaz.syntax.monad._
  import scalaz.syntax.traverse0._

and then use both Monad and Traverse syntax extensions without ambiguity.

For some examples, see the diff of SyntaxUsage.scala.

IMPLEMENTATION COMMENTS:

Where previously we had

  trait ToTraverseOps extends ToFunctorOps with ToFoldableOps

we now have (to be refined shortly)

  trait ToTraverseOps0

  trait ToTraverseOps extends ToTraverseOps0 with ToFunctorOps with ToFoldableOps

That is, we provide ToTraverseOps0 which doesn't inherit Functor and Foldable ops.

In order for the inherited ToFunctorOps and ToFoldableOps to actually require
a Traverse instance to work, we pass the required typeclass as a type argument.
Thus we have

  trait ToTraverseOps0[TC[F[_]] <: Traverse[F]]

  trait ToTraverseOps[TC[F[_]] <: Traverse[F]] extends ToTraverseOps0[TC] with ToFunctorOps[TC] with ToFoldableOps[TC]

and then in ToFunctorOps, we require an instance of TC instead of Functor:

  trait ToFunctorOps0[TC[F[_]] <: Functor[F]] {
    implicit def ToFunctorOps[F[_], A](v: F[A])(implicit F0: TC[F]) = ???
  }

There is one more point worth mentioning.
Consider syntax for MonadTell extending syntax for Monad, using this scheme:

  trait ToMonadOps[TC[F[_]] <: Monad[F]]

  trait ToMonadTellOps[TC[F[_], S] <: MonadTell[F, S]] extends ToMonadOps[???]

The kinds of type arguments of ToMonadOps and ToMonadTellOps don't match.
Therefore, we cannot just pass the `TC` type parameter from ToMonadTellOps
to ToMonadOps. The solution is to use an existential

  trait ToMonadTellOps[TC[F[_], S] <: MonadTell[F, S]] extends ToMonadOps[λ[F[_] => TC[F, S] forSome { type S }]]

and, despite many problems with existentials in Scala, the compiler does
consider MonadTell[F, S] forSome { type S } to be a subtype of Monad[F].