Modernize meta.Ops, adapt it to new useful shapes, and use it more.
Since 8.0.x won't support scala 2.10, there's no reason to bother
using raw trees and `c.Expr`s all over the place. With this, it's
possible to drop all the type parameters and write the impls in a
somewhat less esoteric style (for when more shapes are inevitably
needed).

The ops macros need to be blackbox, otherwise `f_0` won't fix the
result type based on its type argument, and infer `Nothing`. It's
not trivial to return a properly-typed tree if the return type is
an arbitrary function of a cloud of types floating over the macro
application, so it's easier to just accept the blackboxity and go
ahead with unrestricted type inference.

All this, and now it fixes #1542.