Add `*Functions` traits for all typeclasses, mixed into `Prelude`.
No exceptions here.

Methods like `pure`, which do not allow for `F` to be inferred from
the argument types, also get a `*0` variant that uses the "partially
applied" hack. A few methods are also exported as natural transformations,
named `*NT`, as I saw fit.

`*Functions` traits have been set up for all of the subpackages that
contain typeclasses, and *those* now get mixed into `Prelude`, to keep
it shorter.

I have used `F` and `A` as type parameters throughout, irrespective
of the type parameters on the actual typeclass. We should sort that out.