Benchmark tests (#1992)
If you'd like to benchmark the tests, this is a good way. We make sure to fork 5 JVMs and only actually run the benchmark once per iteration for realistic conditions (no JIT, classloading has to be redone on each iteration). The best way to use this is with async-profiler, as described at https://github.com/ktoso/sbt-jmh.