Make @@ abstract and add implicit scope to it
One of the most important use cases of tagged types is to define
alternative typeclass instances. But the fact that currently the
companion objects of the tag and the tagged type are not in the
implicit scope makes the whole feature very cumbersome to use.
We rectify the situation by adding implicit scope to tagged types.